# Allow `Real` arguments in R functions and use Julia implementations only