--- a/src/StatsFuns.jl
+++ b/src/StatsFuns.jl
@@ -35,6 +35,7 @@ import ChainRulesCore
 @reexport using LogExpFunctions:
     xlogx,          # x * log(x) for x > 0, or 0 when x == 0
     xlogy,          # x * log(y) for x > 0, or 0 when x == 0
+    xlog1py,        # x * log(1 + y) for x > 0, or 0 when x == 0
     logistic,       # 1 / (1 + exp(-x))
     logit,          # log(x / (1 - x))
     log1psq,        # log(1 + x^2)