--- a/src/distrs/binom.jl
+++ b/src/distrs/binom.jl
@@ -18,6 +18,7 @@ binompdf(n::Real, p::Real, k::Real) = exp(binomlogpdf(n, p, k))
 
 binomlogpdf(n::Real, p::Real, k::Real) = binomlogpdf(promote(n, p, k)...)
 function binomlogpdf(n::T, p::T, k::T) where {T<:Real}
-    val = betalogpdf(k + 1, n - k + 1, p) - log(n + 1)
-    return 0 <= k <= n && isinteger(n) && isinteger(k) ? val : oftype(val, -Inf)
+    m = clamp(k, 0, n)
+    val = betalogpdf(m + 1, n - m + 1, p) - log(n + 1)
+    return 0 <= k <= n && isinteger(k) ? val : oftype(val, -Inf)
 end