--- a/src/distrs/fdist.jl
+++ b/src/distrs/fdist.jl
@@ -1,8 +1,8 @@
 # functions related to F distribution
 
-import .RFunctions:
-    fdistpdf,
-    fdistlogpdf,
+# R implementations
+# For pdf and logpdf we use the Julia implementation
+using .RFunctions:
     fdistcdf,
     fdistccdf,
     fdistlogcdf,
@@ -12,8 +12,14 @@ import .RFunctions:
     fdistinvlogcdf,
     fdistinvlogccdf
 
-# pdf for numbers with generic types
-fdistpdf(ν1::Real, ν2::Real, x::Number) = sqrt((ν1 * x)^ν1 * ν2^ν2 / (ν1 * x + ν2)^(ν1 + ν2)) / (x * beta(ν1 / 2, ν2 / 2))
+# Julia implementations
+fdistpdf(ν1::Real, ν2::Real, x::Real) = exp(fdistlogpdf(ν1, ν2, x))
 
-# logpdf for numbers with generic types
-fdistlogpdf(ν1::Real, ν2::Real, x::Number) = (ν1 * log(ν1 * x) + ν2 * log(ν2) - (ν1 + ν2) * log(ν1 * x + ν2)) / 2 - log(x) - logbeta(ν1 / 2, ν2 / 2)
+fdistlogpdf(ν1::Real, ν2::Real, x::Real) = fdistlogpdf(promote(ν1, ν2, x)...)
+function fdistlogpdf(ν1::T, ν2::T, x::T) where {T<:Real}
+    # we ensure that `log(x)` does not error if `x < 0`
+    ν1ν2 = ν1 / ν2
+    y = max(x, 0)
+    val = (xlogy(ν1, ν1ν2) + xlogy(ν1 - 2, y) - xlogy(ν1 + ν2, 1 + ν1ν2 * y)) / 2 - logbeta(ν1 / 2, ν2 / 2)
+    return x < 0 ? oftype(val, -Inf) : val
+end