--- a/src/distrs/nbinom.jl
+++ b/src/distrs/nbinom.jl
@@ -1,6 +1,7 @@
 # functions related to negative binomial distribution
 
-import .RFunctions:
+# R implementations
+using .RFunctions:
     nbinompdf,
     nbinomlogpdf,
     nbinomcdf,