--- a/src/distrs/nchisq.jl
+++ b/src/distrs/nchisq.jl
@@ -1,6 +1,7 @@
 # functions related to noncentral chi-square distribution
 
-import .RFunctions:
+# R implementations
+using .RFunctions:
     nchisqpdf,
     nchisqlogpdf,
     nchisqcdf,