--- a/src/distrs/pois.jl
+++ b/src/distrs/pois.jl
@@ -16,4 +16,7 @@ using .RFunctions:
 poispdf(λ::Real, x::Real) = exp(poislogpdf(λ, x))
 
 poislogpdf(λ::Real, x::Real) = poislogpdf(promote(λ, x)...)
-poislogpdf(λ::T, x::T) where {T <: Real} = xlogy(x, λ) - λ - loggamma(x + 1)
+function poislogpdf(λ::T, x::T) where {T <: Real}
+    val = xlogy(x, λ) - λ - loggamma(x + 1)
+    return x >= 0 && isinteger(x) ? val : oftype(val, -Inf)
+end