--- a/src/rmath.jl
+++ b/src/rmath.jl
@@ -6,18 +6,7 @@ import Rmath: libRmath
 
 ### import macro
 
-# union of base types that can be safely converted to Float64
-# we do not use `Real` generally since for some distribution
-# functions a fallback Julia implementation for `Real`s is provided
-# TODO: switch to Julia implementation whenever possible and numerically accurate
-const IntOrRationalOrFloat = Union{
-    Int8,Int16,Int32,Int64,UInt8,UInt16,UInt32,
-    Rational{Int8},Rational{Int16},Rational{Int32},Rational{Int64},
-    Rational{UInt8},Rational{UInt16},Rational{UInt32},Rational{UInt64},
-    Float16,Float32,Float64,
-}
-
-function _import_rmath(rname::Symbol, jname::Symbol, pargs, T::Symbol=:IntOrRationalOrFloat)
+function _import_rmath(rname::Symbol, jname::Symbol, pargs)
     # C function names
     if rname == :norm
         dfun = Expr(:quote, "dnorm4")
@@ -68,7 +57,7 @@ function _import_rmath(rname::Symbol, jname::Symbol, pargs, T::Symbol=:IntOrRati
         rtypes = Expr(:tuple, _pts...)
     end
 
-    pdecls = [Expr(:(::), ps, T) for ps in pargs] # [:(p1::T), :(p2::T), ...]
+    pdecls = [Expr(:(::), ps, :Real) for ps in pargs] # [:(p1::Real), :(p2::Real), ...]
 
     if is_tukey
         # ptukey and qtukey have an extra literal 1 argument
@@ -78,53 +67,53 @@ function _import_rmath(rname::Symbol, jname::Symbol, pargs, T::Symbol=:IntOrRati
     # Function implementation
     quote
         if $(!is_tukey)
-            function $pdf($(pdecls...), x::$T)
+            function $pdf($(pdecls...), x::Real)
                 T = float(Base.promote_typeof($(pargs...), x))
                 return convert(T, ccall(($dfun, libRmath), Float64, $dtypes, x, $(pargs...), 0))
             end
 
-            function $logpdf($(pdecls...), x::$T)
+            function $logpdf($(pdecls...), x::Real)
                 T = float(Base.promote_typeof($(pargs...), x))
                 return convert(T, ccall(($dfun, libRmath), Float64, $dtypes, x, $(pargs...), 1))
             end
         end
 
-        function $cdf($(pdecls...), x::$T)
+        function $cdf($(pdecls...), x::Real)
             T = float(Base.promote_typeof($(pargs...), x))
             return convert(T, ccall(($pfun, libRmath), Float64, $ptypes, x, $(pargs...), 1, 0))
         end
 
-        function $ccdf($(pdecls...), x::$T)
+        function $ccdf($(pdecls...), x::Real)
             T = float(Base.promote_typeof($(pargs...), x))
             return convert(T, ccall(($pfun, libRmath), Float64, $ptypes, x, $(pargs...), 0, 0))
         end
 
-        function $logcdf($(pdecls...), x::$T)
+        function $logcdf($(pdecls...), x::Real)
             T = float(Base.promote_typeof($(pargs...), x))
             return convert(T, ccall(($pfun, libRmath), Float64, $ptypes, x, $(pargs...), 1, 1))
         end
 
-        function $logccdf($(pdecls...), x::$T)
+        function $logccdf($(pdecls...), x::Real)
             T = float(Base.promote_typeof($(pargs...), x))
             return convert(T, ccall(($pfun, libRmath), Float64, $ptypes, x, $(pargs...), 0, 1))
         end
 
-        function $invcdf($(pdecls...), q::$T)
+        function $invcdf($(pdecls...), q::Real)
             T = float(Base.promote_typeof($(pargs...), q))
             return convert(T, ccall(($qfun, libRmath), Float64, $qtypes, q, $(pargs...), 1, 0))
         end
 
-        function $invccdf($(pdecls...), q::$T)
+        function $invccdf($(pdecls...), q::Real)
             T = float(Base.promote_typeof($(pargs...), q))
             return convert(T, ccall(($qfun, libRmath), Float64, $qtypes, q, $(pargs...), 0, 0))
         end
 
-        function $invlogcdf($(pdecls...), lq::$T)
+        function $invlogcdf($(pdecls...), lq::Real)
             T = float(Base.promote_typeof($(pargs...), lq))
             return convert(T, ccall(($qfun, libRmath), Float64, $qtypes, lq, $(pargs...), 1, 1))
         end
 
-        function $invlogccdf($(pdecls...), lq::$T)
+        function $invlogccdf($(pdecls...), lq::Real)
             T = float(Base.promote_typeof($(pargs...), lq))
             return convert(T, ccall(($qfun, libRmath), Float64, $qtypes, lq, $(pargs...), 0, 1))
         end
@@ -140,10 +129,6 @@ macro import_rmath(rname, jname, pargs...)
     esc(_import_rmath(rname, jname, pargs))
 end
 
-macro import_rmath_real(rname, jname, pargs...)
-    esc(_import_rmath(rname, jname, pargs, :Real))
-end
-
 ### Import specific functions
 
 @import_rmath beta beta α β
@@ -151,16 +136,16 @@ end
 @import_rmath chisq chisq k
 @import_rmath f fdist ν1 ν2
 @import_rmath gamma gamma k θ
-@import_rmath_real hyper hyper ms mf n
+@import_rmath hyper hyper ms mf n
 @import_rmath norm norm μ σ
-@import_rmath_real nbinom nbinom r p
+@import_rmath nbinom nbinom r p
 @import_rmath pois pois λ
 @import_rmath t tdist k
-@import_rmath_real tukey srdist k ν
+@import_rmath tukey srdist k ν
 
-@import_rmath_real nbeta nbeta α β λ
-@import_rmath_real nchisq nchisq k λ
-@import_rmath_real nf nfdist k1 k2 λ
-@import_rmath_real nt ntdist k λ
+@import_rmath nbeta nbeta α β λ
+@import_rmath nchisq nchisq k λ
+@import_rmath nf nfdist k1 k2 λ
+@import_rmath nt ntdist k λ
 
 end