--- a/test/chainrules.jl
+++ b/test/chainrules.jl
@@ -32,25 +32,22 @@ using Random
     test_frule(tdistlogpdf, x, y)
     test_rrule(tdistlogpdf, x, y)
 
-    # use `BigFloat` to avoid Rmath implementation in finite differencing check
-    # (returns `NaN` for non-integer values)
-    n = rand(1:100)
-    x = BigFloat(n)
-    y = big(logistic(randn()))
-    z = BigFloat(rand(1:n))
+    x = rand(1:100)
+    y = logistic(randn())
+    z = rand(1:x)
     test_frule(binomlogpdf, x, y, z)
     test_rrule(binomlogpdf, x, y, z)
 
-    x = big(exp(randn()))
-    y = BigFloat(rand(1:100))
+    x = exp(randn())
+    y = rand(1:100)
     test_frule(poislogpdf, x, y)
     test_rrule(poislogpdf, x, y)
 
     # test special case λ = 0
-    _, pb = rrule(StatsFuns.poislogpdf, 0.0, 0.0)
+    _, pb = rrule(poislogpdf, 0.0, 0)
     _, x̄1, _ = pb(1)
     @test x̄1 == -1
-    _, pb = rrule(StatsFuns.poislogpdf, 0.0, 1.0)
+    _, pb = rrule(poislogpdf, 0.0, 1)
     _, x̄1, _ = pb(1)
     @test x̄1 == Inf
 end