@devmotion 2021-09-13
Bump :slightly_smiling_face: 

 It would also be nice to merge this PR since it fixes the remaining test errors in https://github.com/JuliaStats/Distributions.jl/issues/1387.